Migration 4 Change datatype from int into bigint

+++ Database/dbo/Tables/CheckpointPasses.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[CheckpointPasses] (
     [RacerId]              INT           NOT NULL,
-    [Id]                   INT           IDENTITY (1, 1) NOT NULL,
+    [Id]                   BIGINT        IDENTITY (1, 1) NOT NULL,
     [CheckpointId]         INT           NOT NULL,
     [Time]                 DATETIME2 (7) NOT NULL,
     [DistanceToCheckPoint] FLOAT (53)    CONSTRAINT [DF_CheckpointPasses_DistanceToCheckPoint] DEFAULT ((0)) NOT NULL,
@@ -8,6 +8,8 @@
     CONSTRAINT [FK_Checkpointpasses_Checkpoints] FOREIGN KEY ([CheckpointId]) REFERENCES [dbo].[Checkpoints] ([Id]) ON DELETE CASCADE ON UPDATE CASCADE,
     CONSTRAINT [FK_Checkpointpasses_Racers] FOREIGN KEY ([RacerId]) REFERENCES [dbo].[Racers] ([Id]) ON DELETE CASCADE ON UPDATE CASCADE
 ) ON [PS_RACER_BY_ID] ([RacerId]);
+
+
